Use HandlerKeys directly without casting

+++ src/kek.cpp
@@ -23,12 +23,9 @@ void LaunchApp() {
   x_window->show();
 
   EventHandler ehandler;
-  ehandler.add_mouse_motion_listener(static_cast<int>(constants::HandlerKeys::WINDOW), 
-      x_window);
-  ehandler.add_mouse_motion_listener(static_cast<int>(constants::HandlerKeys::WINDOW_VIEW), 
-      x_window->view);
-  ehandler.add_mouse_button_press_listener(static_cast<int>(constants::HandlerKeys::WINDOW_VIEW), 
-      x_window->view);
+  ehandler.add_mouse_motion_listener(constants::HandlerKeys::WINDOW, x_window);
+  ehandler.add_mouse_motion_listener(constants::HandlerKeys::WINDOW_VIEW, x_window->view);
+  ehandler.add_mouse_button_press_listener(constants::HandlerKeys::WINDOW_VIEW, x_window->view);
   ehandler.event_handler_loop(*x_window);
 }
